This commit will: 1. Add a debug config to testing configs 2. Add scaling to tests
